PluginClassLoader and PluginManager impl do not belong into the API module, moved to proxy module
