+++ api/src/main/java/net/md_5/bungee/api/plugin/PluginClassLoader.java
@@ -1,0 +1,6 @@
+package net.md_5.bungee.api.plugin;
+
+interface PluginClassLoader
+{
+    void init(Plugin plugin);
+}
